+++ lib/bloblang/x/parser/parser.go
@@ -518,7 +518,7 @@ func Comment() Type {
 		Sequence(
 			Char('#'),
 			JoinStringPayloads(
-				UntilFail(NotChar('\n')),
+				Optional(UntilFail(NotChar('\n'))),
 			),
 			Newline(),
 		),